fix(mint): snackbar on creating
